need to fix printed values

+++ gazeplay-commons/src/main/java/net/gazeplay/commons/utils/FixationPoint.java
@@ -7,10 +7,10 @@ public class FixationPoint {
 
     @Setter
     @Getter
-    private int firstGaze;
+    private long firstGaze;
     @Setter
     @Getter
-    private int gazeDuration;
+    private long gazeDuration;
 
     public FixationPoint() {
         firstGaze = 0;
